refactor: extract path parameter interpolation logic into a new utility function for improved modularity and clarity
